Add first step for visitor sort on userlist
Decided to show an empty disabled option if the webmaster has not set an initial sorting on the list.
The sort options are still hardcoded in the JS, not dynamic yet.
The sorting itself is also not implemented yet.
Also added filter label to keep filter and sort elements more consistent.