Fix userlist when name property is not there
This might happen when no fields are set to 'Map result', in which case Cyclos returns the user's name in a 'display' property. The plugin will now show that field if it exists. If some fields are set to 'Map result', but not the Full name, Cyclos returns no name at all, nor in a name property nor in a display property. The plugin now simple shows nothing for the name in that case, instead of throwing an error.